refactor: (#508) silentLogin 함수를 유틸 파일로 분리
