+++ frontend/src/utils/fetch.ts
@@ -1,7 +1,5 @@
-import { postTokens } from '@api/token';
-
-import { clearCookieToken, getCookieToken, setCookieToken } from './cookie';
-import { isExpiredAccessToken } from './isExpiredAccessToken';
+import { getCookieToken } from './cookie';
+import { silentLogin } from './silentLogin';
 
 const headers = {
   Authorization: `Bearer `,
@@ -23,30 +21,6 @@ const makeFetchMultiHeaders = () => {
   return {
     Authorization: `Bearer ${cookie.accessToken}`,
   };
-};
-
-export const silentLogin = async () => {
-  const refreshToken = getCookieToken().refreshToken;
-
-  if (!refreshToken || !isExpiredAccessToken()) {
-    return;
-  }
-
-  try {
-    const tokenData = await postTokens(refreshToken);
-
-    const updatedAccessToken = tokenData.accessToken;
-    const updatedRefreshToken = tokenData.refreshToken;
-
-    setCookieToken('accessToken', updatedAccessToken);
-    setCookieToken('refreshToken', updatedRefreshToken);
-  } catch (error) {
-    clearCookieToken('accessToken');
-    clearCookieToken('refreshToken');
-    window.location.href = '/login';
-
-    throw new Error('로그인에 실패했습니다. 다시 로그인 해주세요.');
-  }
 };
 
 export const getFetch = async <T>(url: string): Promise<T> => {
